requirements.txt
asgiref==3.5.2
autopep8==2.0.0
Django==4.1.2
graphql-core==3.2.3
pycodestyle==2.9.1
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.3
strawberry-graphql==0.138.1
strawberry-graphql-django==0.7.1
tomli==2.0.1
typing_extensions==4.4.0
